Add ratelimits to messaging and voting
These use the ratelimit library so they allow for burstiness. We have to
live configurables: the length of the burst window (in seconds) and the
number of events per second we'll allow through on average during that
window.

This does not add any frontend display of a ratelimit response, but does
pass back information which could be used to do so (i.e. how much time
until the ratelimit resets).